requirements.txt
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
openai==1.6.1
python-dotenv==1.0.0gunicorn==21.2.0
